+++ crates/macro-support/src/codegen.rs
@@ -2319,7 +2319,7 @@ impl TryToTokens for DescribeImport<'_> {
 impl ToTokens for ast::Enum {
     fn to_tokens(&self, into: &mut TokenStream) {
         let enum_name = &self.rust_name;
-        let name_str = self.js_name.to_string();
+        let name_str = shared::qualified_name(self.js_namespace.as_deref(), &self.js_name);
         let name_len = name_str.len() as u32;
         let name_chars = name_str.chars().map(|c| c as u32);
         let hole = &self.hole;